implementation(profile): fixing staging file crash and added UI for edit photo
