+++ app/build.gradle
@@ -66,17 +66,18 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     testImplementation "androidx.arch.core:core-testing:2.2.0" // InstantTaskExecutorRule
-    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.4" //TestDispatcher
+    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.7.1" //TestDispatcher
     androidTestImplementation 'androidx.arch.core:core-testing:2.2.0'
-    androidTestImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.4'
+    androidTestImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.7.1'
 
     //paging
     implementation "androidx.paging:paging-runtime-ktx:3.1.1"
+    implementation "androidx.paging:paging-compose:1.0.0-alpha20"
  
     // camera
-    implementation "androidx.camera:camera-camera2:1.2.2"
-    implementation "androidx.camera:camera-lifecycle:1.2.2"
-    implementation "androidx.camera:camera-view:1.2.2"
+    implementation "androidx.camera:camera-camera2:1.2.3"
+    implementation "androidx.camera:camera-lifecycle:1.2.3"
+    implementation "androidx.camera:camera-view:1.2.3"
 
     // retrofit
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
