--- app/src/main/java/com/example/bangkitandroid/data/Repository.kt
@@ -1,62 +1,0 @@
-package com.example.bangkitandroid.data
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MediatorLiveData
-import com.example.bangkitandroid.data.remote.retrofit.ApiService
-import com.example.bangkitandroid.domain.entities.Blog
-import com.example.bangkitandroid.domain.entities.Disease
-import com.example.bangkitandroid.domain.entities.User
-import com.example.bangkitandroid.service.DummyData
-import com.example.bangkitandroid.service.Result
-
-class Repository (
-    private val apiService: ApiService,
-) {
-    private val historyResult = MediatorLiveData<Result<List<Disease>>>()
-    private val blogResult = MediatorLiveData<Result<List<Blog>>>()
-    private val getUserResult = MediatorLiveData<Result<User>>()
-    private val editProfileResult = MediatorLiveData<Result<User>>()
-    private val loginResult = MediatorLiveData<Result<User>>()
-    private val registerResult = MediatorLiveData<Result<User>>()
-    
-    fun login(phoneNumber: String, password: String): LiveData<Result<User>> {
-        loginResult.value = Result.Success(DummyData().getUserDummy(1))
-        return loginResult
-    }
-    
-    fun register(name: String, phoneNumber: String, password: String): LiveData<Result<User>> {
-        registerResult.value = Result.Success(DummyData().getUserDummy(1))
-        return registerResult
-    }
-
-    fun getHistory(token: String): LiveData<Result<List<Disease>>> {
-        historyResult.value = Result.Success(DummyData().getHistoryDiseasesDummy())
-        return historyResult
-    }
-
-    fun getListBlog(): LiveData<Result<List<Blog>>> {
-        blogResult.value = Result.Success(DummyData().getListBlogsDummy())
-        return blogResult
-    }
-
-    fun getUser(): LiveData<Result<User>> {
-        getUserResult.value = Result.Success(DummyData().getUserDummy(1))
-        return getUserResult
-    }
-
-    fun editProfile(name: String, phoneNumber: String): LiveData<Result<User>> {
-        editProfileResult.value = Result.Success(DummyData().getUserDummy(1))
-        return editProfileResult
-    }
-
-    companion object {
-        @Volatile
-        private var instance: Repository? = null
-        fun getInstance(
-            apiService: ApiService,
-        ): Repository = instance ?: synchronized(this){
-            instance ?: Repository(apiService)
-        }.also {
-            instance = it
-        }
-    }
-}
