+++ app/src/main/java/com/example/bangkitandroid/data/remote/retrofit/ApiService.kt
@@ -1,50 +1,29 @@
 package com.example.bangkitandroid.data.remote.retrofit
-import com.example.bangkitandroid.data.remote.response.DiseaseHistoryResponse
-import com.example.bangkitandroid.data.remote.response.ListBlogResponse
-import com.example.bangkitandroid.data.remote.response.DiseaseResponse
-import com.example.bangkitandroid.data.remote.response.EditProfileResponse
+
 import okhttp3.MultipartBody
-import retrofit2.Call
 import retrofit2.http.*
 import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.Query
 import retrofit2.http.Field
 import retrofit2.http.FormUrlEncoded
-import retrofit2.http.Header
 import retrofit2.http.POST
-
+import retrofit2.http.Multipart
+import retrofit2.http.Part
 import com.example.bangkitandroid.data.remote.response.LoginResponse
 import com.example.bangkitandroid.data.remote.response.RegisterResponse
-import retrofit2.http.Field
-import retrofit2.http.FormUrlEncoded
-import retrofit2.http.POST
-
 import com.example.bangkitandroid.data.remote.response.BlogResponse
 import com.example.bangkitandroid.data.remote.response.CommentResponse
+import com.example.bangkitandroid.data.remote.response.DiseaseHistoryResponse
+import com.example.bangkitandroid.data.remote.response.DiseaseResponse
+import com.example.bangkitandroid.data.remote.response.EditProfileResponse
 import com.example.bangkitandroid.data.remote.response.ListBlogResponse
-import okhttp3.MultipartBody
-import retrofit2.http.Field
-import retrofit2.http.FormUrlEncoded
-import retrofit2.http.GET
-import retrofit2.http.Header
-import retrofit2.http.Multipart
-import retrofit2.http.POST
-import retrofit2.http.Part
-import retrofit2.http.Query
 
-interface ApiService 
+interface ApiService {
     @GET("blog")
     fun getBlog(
         @Query("id") id: Int,
     ): BlogResponse
-
-    @GET("history")
-    fun getDiseases(
-        @Header("Authorization") token: String,
-        @Query("Page") page: Int,
-        @Query("Size") size: Int,
-    ): DiseaseHistoryResponse
 
     @GET("blogs")
     fun getBlogs(
