+++ app/src/main/java/com/example/bangkitandroid/service/Injection.kt
@@ -2,7 +2,6 @@ package com.example.bangkitandroid.service
 
 import android.content.Context
 import com.example.bangkitandroid.data.remote.Repository
-import com.example.bangkitandroid.data.Repository
 import com.example.bangkitandroid.data.remote.retrofit.ApiConfig
 
 object Injection {
