+++ app/src/main/java/com/example/bangkitandroid/ui/authentication/RegisterActivity.kt
@@ -2,9 +2,6 @@ package com.example.bangkitandroid.ui.authentication
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import com.example.bangkitandroid.R
-import com.example.bangkitandroid.databinding.ActivityDiseaseHistoryBinding
-import com.example.bangkitandroid.databinding.ActivityLoginBinding
 import com.example.bangkitandroid.databinding.ActivityRegisterBinding
 
 class RegisterActivity : AppCompatActivity() {
