+++ app/src/main/java/com/example/bangkitandroid/ui/home/HomeViewModel.kt
@@ -1,10 +1,10 @@
 package com.example.bangkitandroid.ui.home
 
 import androidx.lifecycle.ViewModel
-import com.example.bangkitandroid.data.Repository
+import com.example.bangkitandroid.data.remote.Repository
 
 class HomeViewModel(private val repository: Repository): ViewModel() {
     fun getHistory(token: String) = repository.getHistory(token)
 
-    fun getBlog() = repository.getListBlog()
+    fun getBlog() = repository.getListBlogHome()
 }
