+++ app/src/main/java/com/example/bangkitandroid/ui/profile/ProfileViewModel.kt
@@ -1,7 +1,7 @@
 package com.example.bangkitandroid.ui.profile
 
 import androidx.lifecycle.ViewModel
-import com.example.bangkitandroid.data.Repository
+import com.example.bangkitandroid.data.remote.Repository
 
 class ProfileViewModel(private val repository: Repository) : ViewModel() {
     fun getUser() = repository.getUser()
