+++ app/src/main/res/layout/activity_profile_logged.xml
@@ -19,7 +19,7 @@
         android:layout_marginBottom="16dp"
         android:textColor="@color/black"
         android:drawablePadding="4dp"
-        app:drawableStartCompat="@drawable/leaf" />
+        app:drawableStartCompat="@drawable/ic_leaf" />
 
     <TextView
         android:id="@+id/edit_tv"
