+++ app/src/main/res/values/strings.xml
@@ -30,7 +30,7 @@
     <string name="rekomendasi_produk">Rekomendasi Produk</string>
     <string name="back_button">Back Button</string>
     <string name="unggah_gambar">Unggah Gambar</string>
-    <string name="ambil_ulang">ambil ulang</string>
+    <string name="ambil_ulang">Ambil ulang</string>
     <string name="image_preview">Image Preview</string>
     <string name="sedang_diproses">Sedang Diproses</string>
     <string name="judul_blog">Judul Blog</string>
@@ -46,4 +46,6 @@
     <string name="masukkan_nama">Masukkan Nama</string>
     <string name="nama">Nama</string>
     <string name="daftar">Daftar</string>
+    <string name="take_photo">Ambil foto</string>
+    <string name="gallery">Galeri</string>
 </resources>
