+++ app/src/test/java/com/example/bangkitandroid/ui/authentication/AuthenticationViewModelTest.kt
@@ -1,7 +1,7 @@
 package com.example.bangkitandroid.ui.authentication
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
-import com.example.bangkitandroid.data.Repository
+import com.example.bangkitandroid.data.remote.Repository
 import com.example.bangkitandroid.data.remote.retrofit.ApiService
 import com.example.bangkitandroid.service.DummyData
 import com.example.bangkitandroid.service.Result
