+++ app/src/test/java/com/example/bangkitandroid/ui/home/HomeViewModelTest.kt
@@ -1,13 +1,13 @@
 package com.example.bangkitandroid.ui.home
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
-import com.example.bangkitandroid.data.Repository
+import com.example.bangkitandroid.data.remote.Repository
 import com.example.bangkitandroid.data.remote.retrofit.ApiService
 import com.example.bangkitandroid.domain.entities.Blog
 import com.example.bangkitandroid.domain.entities.Disease
 import com.example.bangkitandroid.service.DummyData
 import com.example.bangkitandroid.service.Result
-import com.example.bangkitandroid.utils.FakeApiService
+import com.example.bangkitandroid.data.remote.FakeApiService
 import com.example.bangkitandroid.utils.MainDispatcherRule
 import com.example.bangkitandroid.utils.observeForTesting
 import kotlinx.coroutines.ExperimentalCoroutinesApi
