+++ app/src/test/java/com/example/bangkitandroid/ui/profile/ProfileViewModelTest.kt
@@ -1,12 +1,12 @@
 package com.example.bangkitandroid.ui.profile
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
-import com.example.bangkitandroid.data.Repository
+import com.example.bangkitandroid.data.remote.Repository
 import com.example.bangkitandroid.data.remote.retrofit.ApiService
 import com.example.bangkitandroid.domain.entities.User
 import com.example.bangkitandroid.service.DummyData
 import com.example.bangkitandroid.service.Result
-import com.example.bangkitandroid.utils.FakeApiService
+import com.example.bangkitandroid.data.remote.FakeApiService
 import com.example.bangkitandroid.utils.MainDispatcherRule
 import com.example.bangkitandroid.utils.observeForTesting
 import kotlinx.coroutines.ExperimentalCoroutinesApi
